Move document serializers into a submodule.
We'll split up this file next.
